// src/models/PerfilState.ts
export interface PerfilState {
    idDono: number
    nomePerfil: string
    usuarioLogado: number
    logadoSSO: number
    multiLiberado: boolean
    permissoes: any
    listaUsuarios: any
    usouDiasGratis: boolean
    planoAtivo: boolean,
    carregado: boolean
    permiteTabelasPersonalizadas: boolean
}